Move FIPS test fixtures to conftest and share them.
